+++ test/only_necessary_derives/models/todos/generated.rs
@@ -8,7 +8,7 @@ use serde::{Deserialize, Serialize};
 
 type Connection = diesel::r2d2::PooledConnection<diesel::r2d2::ConnectionManager<diesel::PgConnection>>;
 
-#[derive(Debug, Serialize, Deserialize, Clone, Queryable, AsChangeset, Selectable)]
+#[derive(Debug, Clone, Serialize, Deserialize, Queryable, Selectable, AsChangeset)]
 #[diesel(table_name=todos, primary_key(id))]
 pub struct Todos {
     pub id: i32,
@@ -16,14 +16,14 @@ pub struct Todos {
     pub created_at: chrono::NaiveDateTime,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone, Insertable, AsChangeset)]
+#[derive(Debug, Clone, Serialize, Deserialize, Insertable, AsChangeset)]
 #[diesel(table_name=todos)]
 pub struct CreateTodos {
     pub id: i32,
     pub test: String,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone, AsChangeset)]
+#[derive(Debug, Clone, Serialize, Deserialize, AsChangeset)]
 #[diesel(table_name=todos)]
 pub struct UpdateTodos {
     pub test: Option<String>,